Functional spell select list for both classes
