+++ Assets/Scripts/Controllers/Player/PlayerMovementScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 944fb22c8890d7841bda26a985958d15
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
