+++ Assets/Scripts/Controllers/Player/Warrior/MeleeController.cs
@@ -9,7 +9,6 @@ public class MeleeController : MonoBehaviour
     private AnimationScriptControllerWarrior animations;
     private Sword sword;
 
-    public OverlayController overlayController;
     public AttackIndicator indicator;
     public float attackDelay;
     [HideInInspector]
@@ -43,37 +42,24 @@ public class MeleeController : MonoBehaviour
 
     void FixedUpdate()
     {
-        if (controls.menu && !overlayController.isEnabled)
+        if (controls.mousePressed_1)
         {
-            overlayController.Enable(true);
-            canHit = true;
-        }
-        else if (!controls.menu && overlayController.isEnabled)
-        {
-            overlayController.Enable(false);
-        }
-
-        if (!overlayController.isEnabled)
-        {
-            if (controls.mousePressed_1)
+            if (canHit && !comboLock)
             {
-                if (canHit && !comboLock)
+                if (comboQueue.Count < 3)
                 {
-                    if (comboQueue.Count < 3)
-                    {
 
-                        AttackAnimations();
-                        comboQueue.Add(0);
-                        reset = 0f;
+                    AttackAnimations();
+                    comboQueue.Add(0);
+                    reset = 0f;
 
-                    }
-                    canHit = false;
                 }
+                canHit = false;
             }
-            else
-            {
-                canHit = true;
-            }
+        }
+        else
+        {
+            canHit = true;
         }
 
         if (comboQueue.Count != 0 && !attacking)
