+++ Assets/Scripts/Controllers/Player/Warrior/PlayerMovementScriptWarrior.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using UnityEditor;
 using UnityEngine;
 
-public class PlayerMovementScript : MonoBehaviour
+public class PlayerMovementScriptWarrior : PlayerMovementScript
 {
     public CharacterController controller;
     public Transform indicatorWheel;
@@ -16,16 +16,11 @@ public class PlayerMovementScript : MonoBehaviour
     public float gravity = -9.81f;
     public float groundDistance = 0.4f;
     public float jumpHeight = 2f;
+    public float rollDistance = 10f;
     public float smoothing = 0.1f;
     float smoothVelocity;
     public float runspeed = 0f;
-
-    public ParticleSystem dodgeParticles;
-    public float dodgeDuration = 0.5f;
-    public float dodgeDistance = 10f;
-    public float dodgeCooldown = 0.5f;
-    public float stunAfterDodge = 0.05f;
-
+  
     public Vector3 direction;
    
     Vector3 velocity;
@@ -41,42 +36,51 @@ public class PlayerMovementScript : MonoBehaviour
     // Raw inputs
     public bool mouse_1;
     public bool mouse_2;
+    public bool mousePressed_1;
     // ---------------
-    public bool lockMouseInputs;
 
     private float horizontal;
     private float vertical;
     private bool running;
     private bool jump;
-    private bool dodge;
-    private bool dodging;
-    private float lastDodge;
-    private Vector3 dodgeDirection;
-    private ParticleSystem dodgeParticleSystem;
-    private CameraShake cameraShake;
+
+    //Dodge
+    private bool roll;
+    public bool rolling;
+    private float rollTime = 0.9f;
+    public float rollSpeed=15f;
+    private Vector3 rollDirection;
+    private float rollCooldown;
+    private bool canRoll;
+    //Slide(When attacking)
+    public bool sliding;
+ 
+    private float slidingSpeed = 10f;
+
+    //----------------
+    public GameObject particlesOnHit;
+    //temp
+    private MeleeController meleeController;
+    
 
     private void Start()
     {
+        meleeController = GameObject.FindObjectOfType<MeleeController>() as MeleeController;
         canMove = true;
         casting = false;
         mousedown_1 = false;
         mousedown_2 = false;
+        mousePressed_1 = false;
         menu = false;
-        lockMouseInputs = false;
-
+   
         horizontal = 0f;
         vertical = 0f;
         running = false;
         jump = false;
-        dodge = false;
-        dodging = false;
-        lastDodge = Time.time;
 
-        dodgeParticleSystem = Instantiate(dodgeParticles);
-        dodgeParticleSystem.Stop();
-        dodgeParticleSystem.transform.localScale = Vector3.one;
+        //Roll
+        canRoll = true ;
 
-        cameraShake = FindObjectOfType<CameraShake>();
     }
 
     private void Update()
@@ -100,7 +104,7 @@ public class PlayerMovementScript : MonoBehaviour
         if (!menu)
         {
             // Controlled inputs which lock other if one is pressed
-            if (Input.GetMouseButtonDown(0) && !mousedown_2 && !lockMouseInputs)
+            if (Input.GetMouseButtonDown(0) && !mousedown_2)
             {
                 mousedown_1 = true;
             }
@@ -108,7 +112,7 @@ public class PlayerMovementScript : MonoBehaviour
             {
                 mousedown_1 = false;
             }
-            if (Input.GetMouseButtonDown(1) && !mousedown_1 && !lockMouseInputs)
+            if (Input.GetMouseButtonDown(1) && !mousedown_1)
             {
                 mousedown_2 = true;
             }
@@ -118,7 +122,7 @@ public class PlayerMovementScript : MonoBehaviour
             }
 
             // Raw inputs
-            if (Input.GetMouseButtonDown(0) && !lockMouseInputs)
+            if (Input.GetMouseButtonDown(0))
             {
                 mouse_1 = true;
             }
@@ -126,7 +130,7 @@ public class PlayerMovementScript : MonoBehaviour
             {
                 mouse_1 = false;
             }
-            if (Input.GetMouseButtonDown(1) && !lockMouseInputs)
+            if (Input.GetMouseButtonDown(1))
             {
                 mouse_2 = true;
             }
@@ -134,19 +138,26 @@ public class PlayerMovementScript : MonoBehaviour
             {
                 mouse_2 = false;
             }
+            if (Input.GetMouseButton(0))
+            {
+                mousePressed_1 = true;
+            }
+            else
+            {
+                mousePressed_1 = false;
+            }
         }
 
         running = Input.GetKey(KeyCode.LeftShift);
-        jump = Input.GetKey(KeyCode.C);
+        roll = Input.GetKey(KeyCode.Space);
+        jump = Input.GetKey(KeyCode.L);
 
-        if (!dodging)
-            dodge = Input.GetKey(KeyCode.Space);
     }
 
     // Update is called once per frame
     void FixedUpdate()
     {
-        //=================== GROUND CHECK =================== 
+        //===================GROUND CHECK=================== 
         //check if its close to the ground
         isGrounded = Physics.CheckSphere(groundCheck.position, groundDistance, groundMask);
         //Reset the velocity 
@@ -154,64 +165,29 @@ public class PlayerMovementScript : MonoBehaviour
         {
             if (velocity.y <= -20)
             {
-                StartCoroutine(Stun(2f));
+                StartCoroutine(stun(2f));
             }
             velocity.y = -2f;
         }
-        //=================== Dodge =================== 
-        if (dodge && lastDodge + dodgeCooldown <= Time.time)
-        {
-            dodgeDirection = Quaternion.Euler(0, 45, 0) * new Vector3(horizontal, 0f, vertical).normalized;
-            if (dodgeDirection == Vector3.zero)
-                dodgeDirection = transform.forward;
-            dodgeParticleSystem.transform.position = controller.transform.position;
-            dodgeParticleSystem.transform.rotation = Quaternion.LookRotation(dodgeDirection);
 
-            StartCoroutine(DodgeTimer(dodgeDuration));
+        //===================Movement=================== 
+        if (!rolling)
+        {
+            direction = (Quaternion.Euler(0, 45, 0) * new Vector3(horizontal, 0f, vertical).normalized);
         }
 
-        if (dodging)
+       
+        if ( meleeController.isDuringAttack)  // if mouse down OR if already firing basic
         {
-            controller.Move(dodgeDirection * (dodgeDistance / dodgeDuration) * Time.deltaTime);
-            dodgeParticleSystem.transform.position = controller.transform.position;
-        }
-
-
-        //=================== Movement =================== 
-        if (canMove)
-        {
-            //calculate and normalize direction
-            direction = Quaternion.Euler(0, 45, 0) * new Vector3(horizontal, 0f, vertical).normalized;
-          
+            transform.rotation = indicatorWheel.rotation;
         }
         else
         {
-            direction = new Vector3(0f, 0f, 0f);
-        }
-
-
-        if (mousedown_1 || Wand.castingBasic)  // if mouse down OR if already firing basic
-        {
-            casting = true;
-            if (canMove)
-            {
-                transform.rotation = indicatorWheel.rotation;
-            }
-        }
-        else if (mousedown_2)  // if mouse down OR if already channeling
-        {
-            casting = true;
-            if (canMove)
-            {
-                transform.rotation = indicatorWheel.rotation;
-            }
-        }
-        else
-        {
-            casting = false;
-            float targetAngle = Mathf.Atan2(direction.x, direction.z) * Mathf.Rad2Deg;
-            float angle = Mathf.SmoothDampAngle(transform.eulerAngles.y, targetAngle, ref smoothVelocity, smoothing);
-            if (direction != new Vector3(0, 0, 0)) { transform.rotation = Quaternion.Euler(0f, angle, 0f); }
+         
+                float targetAngle = Mathf.Atan2(direction.x, direction.z) * Mathf.Rad2Deg;
+                float angle = Mathf.SmoothDampAngle(transform.eulerAngles.y, targetAngle, ref smoothVelocity, smoothing);
+                if (direction != new Vector3(0, 0, 0)) { transform.rotation = Quaternion.Euler(0f, angle, 0f); }
+           
         }
         
         //Handles Running
@@ -238,10 +214,29 @@ public class PlayerMovementScript : MonoBehaviour
             }
 
         }
-
+     
         //Move player towords direction
-        controller.Move(direction * (speed+ runspeed) * Time.deltaTime);
+        if (roll && !rolling && canMove)
+        {
+           
+            StartCoroutine(PerformRoll(rollTime));
+           
+        }
+
         
+
+        if (rolling|| sliding)
+        {
+            controller.Move(transform.forward * rollSpeed * Time.deltaTime);
+        }
+        else
+        {
+            if (canMove) controller.Move(direction * (speed + runspeed) * Time.deltaTime);
+        }
+           
+       
+            
+       
         if (jump && isGrounded)
         {
             velocity.y = Mathf.Sqrt(jumpHeight * -2f * gravity);
@@ -254,43 +249,31 @@ public class PlayerMovementScript : MonoBehaviour
         controller.Move(velocity * Time.deltaTime);
     }
 
-    private void DodgeEffect(bool enable)
-    {
-        if (dodgeParticles == null)
-            return;
-
-        if (enable)
-        {
-            dodgeParticleSystem.Play();
-            cameraShake.Shake(dodgeDuration / 2f, 1f);
-        }
-        else
-        {
-            dodgeParticleSystem.Stop();
-        }
-    }
-
-    IEnumerator Stun(float second)
+    public IEnumerator stun(float second)
     {
         canMove = false;
         yield return new WaitForSeconds(second);
         canMove = true;
     }
-
-    IEnumerator DodgeTimer(float seconds)
+    private IEnumerator PerformRoll(float second)
     {
-        // Disable controls and enable particles
-        DodgeEffect(true);
-        canMove = false;
-        dodging = true;
-        dodge = false;
-        // Dodge duration
-        yield return new WaitForSeconds(seconds);
-        // Enable controls and disable particles
-        DodgeEffect(false);
-        dodging = false;
-        canMove = true;
-        lastDodge = Time.time;
-        StartCoroutine(Stun(stunAfterDodge));
+        rolling = true;
+        yield return new WaitForSeconds(second);
+        rolling = false;
     }
+   
+    private IEnumerator ResetCooldown(float second)
+    {
+        canRoll = false;
+        yield return new WaitForSeconds(second);
+        canRoll = true;
+    }
+    //public void getHit()
+    //{
+    //    FindObjectOfType<HitStop>().Stop(0.05f);
+    //    FindObjectOfType<CameraShake>().Shake(0.05f);
+    //    Destroy(Instantiate(particlesOnHit, transform.position + new Vector3(0, 1f, 0), transform.rotation), 1f);
+    //    
+    //}
+
 }
