+++ Assets/Scripts/Controllers/UI/Button Containers/QuickbarButton.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.EventSystems;
 
 public class QuickbarButton : ButtonContainer
 {
@@ -8,4 +9,5 @@ public class QuickbarButton : ButtonContainer
     {
         overlayControls.SetSelectedQuickBar(buttonData.quickBarIndex);
     }
+    
 }
