+++ Assets/Scripts/FollowPlayer.cs
@@ -1,25 +1,22 @@
-﻿
-using UnityEngine;
+﻿using UnityEngine;
 
 public class FollowPlayer : MonoBehaviour
 {
-    public Transform playerPosition;
     public Vector3 offset;
     public float smoothSpeed = 2f;
 
+    private Transform playerPosition;
 
-   
-    bool isWalled;
+    private void Start()
+    {
+        playerPosition = FindObjectOfType<PlayerMovementScript>().transform;
+    }
 
     void FixedUpdate()
     {
-       
-
-
         Vector3 targerPosition = playerPosition.position + offset; ;
         Vector3 smoothedPosition = Vector3.Lerp(transform.position, targerPosition, smoothSpeed * Time.deltaTime);
         transform.position = smoothedPosition;
         transform.LookAt(playerPosition);
-
     }
 }
